requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
bcrypt==4.2.1
billiard==4.2.1
black==24.10.0
boto3==1.35.90
botocore==1.35.90
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
flake8==7.1.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jiter==0.8.2
jmespath==1.0.1
kombu==5.4.2
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.58.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.1
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
